libft

+++ libft/ft_strlen.c
@@ -1,8 +1,8 @@
 #include "libft.h"
 
-int	ft_strlen(char *str)
+size_t	ft_strlen(char *str)
 {
-	int	len;
+	size_t	len;
 
 	len = 0;
 	while (str[len])
